Reorg tensor code and unify CPU/CPU Tensor implementation.
